+++ tests/__init__.py
@@ -11,6 +11,7 @@ class CacheTestMixin(object):
         self.assertEqual(1, cache.getsizeof(None))
         self.assertEqual(1, cache.getsizeof(''))
         self.assertEqual(1, cache.getsizeof(0))
+        self.assertTrue(repr(cache).startswith(cache.__class__.__name__))
 
     def test_insert(self):
         cache = self.make_cache(maxsize=2)
@@ -132,57 +133,13 @@ class CacheTestMixin(object):
         self.assertNotIn(2, cache)
 
         with self.assertRaises(ValueError):
-            cache[4] = 4
+            cache[3] = 4
         self.assertEqual(1, len(cache))
         self.assertEqual(3, cache.currsize)
         self.assertEqual(3, cache[3])
 
-
-class LRUCacheTestMixin(CacheTestMixin):
-
-    def test_lru_insert(self):
-        cache = self.make_cache(maxsize=2)
-
-        cache[1] = 1
-        cache[2] = 2
-        cache[3] = 3
-
-        self.assertEqual(len(cache), 2)
-        self.assertEqual(cache[2], 2)
-        self.assertEqual(cache[3], 3)
-        self.assertNotIn(1, cache)
-
-        cache[2]
-        cache[4] = 4
-        self.assertEqual(len(cache), 2)
-        self.assertEqual(cache[2], 2)
-        self.assertEqual(cache[4], 4)
-        self.assertNotIn(3, cache)
-
-        cache[5] = 5
-        self.assertEqual(len(cache), 2)
-        self.assertEqual(cache[4], 4)
-        self.assertEqual(cache[5], 5)
-        self.assertNotIn(2, cache)
-
-    def test_lru_getsizeof(self):
-        cache = self.make_cache(maxsize=3, getsizeof=lambda x: x)
-
-        cache[1] = 1
-        cache[2] = 2
-
-        self.assertEqual(len(cache), 2)
-        self.assertEqual(cache[1], 1)
-        self.assertEqual(cache[2], 2)
-
-        cache[3] = 3
-
-        self.assertEqual(len(cache), 1)
-        self.assertEqual(cache[3], 3)
-        self.assertNotIn(1, cache)
-        self.assertNotIn(2, cache)
-
         with self.assertRaises(ValueError):
             cache[4] = 4
-        self.assertEqual(len(cache), 1)
-        self.assertEqual(cache[3], 3)
+        self.assertEqual(1, len(cache))
+        self.assertEqual(3, cache.currsize)
+        self.assertEqual(3, cache[3])