+++ tests/test_ttlcache.py
@@ -1,6 +1,6 @@
 import unittest
 
-from . import LRUCacheTestMixin
+from . import CacheTestMixin
 from cachetools import TTLCache, ttl_cache
 
 
@@ -14,26 +14,24 @@ def cached_typed(n):
     return n
 
 
-class TTLCacheTest(unittest.TestCase, LRUCacheTestMixin):
+class Timer:
+    def __init__(self):
+        self.__time = 0
 
-    def make_cache(self, maxsize, getsizeof=None):
-        return TTLCache(maxsize, ttl=0, timer=lambda: 0, getsizeof=getsizeof)
+    def __call__(self):
+        return self.__time
 
-    def make_ttl_cache(self, maxsize, ttl):
-        class Timer:
-            def __init__(self):
-                self.__time = 0
+    def inc(self):
+        self.__time = self.__time + 1
 
-            def __call__(self):
-                return self.__time
 
-            def inc(self):
-                self.__time = self.__time + 1
+class TTLCacheTest(unittest.TestCase, CacheTestMixin):
 
-        return TTLCache(maxsize, ttl, timer=Timer())
+    def make_cache(self, maxsize, getsizeof=None, ttl=0):
+        return TTLCache(maxsize, ttl, timer=Timer(), getsizeof=getsizeof)
 
     def test_ttl_insert(self):
-        cache = self.make_ttl_cache(maxsize=2, ttl=2)
+        cache = self.make_cache(maxsize=2, ttl=2)
         self.assertEqual(cache.ttl, 2)
 
         cache[1] = 1
@@ -58,7 +56,7 @@ class TTLCacheTest(unittest.TestCase, LRUCacheTestMixin):
         self.assertEqual(3, cache[3])
 
     def test_ttl_expire(self):
-        cache = self.make_ttl_cache(maxsize=3, ttl=0)
+        cache = self.make_cache(maxsize=3, ttl=0)
         self.assertEqual(cache.ttl, 0)
 
         cache[1] = 1
@@ -91,7 +89,7 @@ class TTLCacheTest(unittest.TestCase, LRUCacheTestMixin):
         self.assertNotIn(3, cache)
 
     def test_ttl_tuple_key(self):
-        cache = self.make_ttl_cache(maxsize=1, ttl=0)
+        cache = self.make_cache(maxsize=1, ttl=0)
 
         cache[(1, 2, 3)] = 42
         self.assertEqual(42, cache[(1, 2, 3)])
@@ -100,6 +98,53 @@ class TTLCacheTest(unittest.TestCase, LRUCacheTestMixin):
             cache[(1, 2, 3)]
         cache.expire()
         self.assertNotIn((1, 2, 3), cache)
+
+    def test_lru_insert(self):
+        cache = self.make_cache(maxsize=2)
+
+        cache[1] = 1
+        cache[2] = 2
+        cache[3] = 3
+
+        self.assertEqual(len(cache), 2)
+        self.assertEqual(cache[2], 2)
+        self.assertEqual(cache[3], 3)
+        self.assertNotIn(1, cache)
+
+        cache[2]
+        cache[4] = 4
+        self.assertEqual(len(cache), 2)
+        self.assertEqual(cache[2], 2)
+        self.assertEqual(cache[4], 4)
+        self.assertNotIn(3, cache)
+
+        cache[5] = 5
+        self.assertEqual(len(cache), 2)
+        self.assertEqual(cache[4], 4)
+        self.assertEqual(cache[5], 5)
+        self.assertNotIn(2, cache)
+
+    def test_lru_getsizeof(self):
+        cache = self.make_cache(maxsize=3, getsizeof=lambda x: x)
+
+        cache[1] = 1
+        cache[2] = 2
+
+        self.assertEqual(len(cache), 2)
+        self.assertEqual(cache[1], 1)
+        self.assertEqual(cache[2], 2)
+
+        cache[3] = 3
+
+        self.assertEqual(len(cache), 1)
+        self.assertEqual(cache[3], 3)
+        self.assertNotIn(1, cache)
+        self.assertNotIn(2, cache)
+
+        with self.assertRaises(ValueError):
+            cache[4] = 4
+        self.assertEqual(len(cache), 1)
+        self.assertEqual(cache[3], 3)
 
     def test_decorator(self):
         self.assertEqual(cached.cache_info(), (0, 0, 2, 0))